requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
pytest>=6.2.5
pyyaml>=5.4.1
python-dotenv>=0.19.0
sqlalchemy>=1.4.23
requests>=2.26.0
joblib>=1.0.1
matplotlib>=3.4.3
seaborn>=0.11.2
pytest-cov>=2.12.1 